Add in caching functionality for mid-level-layers
