+++ src/layers/microscopy-viewer-layer/data-utils/index.js
@@ -1,4 +1,4 @@
-import loadZarr from './zarr-utils';
+import {loadZarr, getZarrConnections} from './zarr-utils';
 import {loadTiff, getTiffConnections} from './tiff-utils';
 
-export {loadZarr, loadTiff, getTiffConnections}
+export {loadZarr, loadTiff, getTiffConnections, getZarrConnections}
