+++ src/layers/microscopy-viewer-layer/data-utils/zarr-utils.js
@@ -1,41 +1,23 @@
 import { slice, openArray } from 'zarr';
 
-var zarrArrays = {}
-
-async function getData({ config, tileSize, x, y, stride, tilingWidth }) {
+async function getData({ arr, channel, tileSize, x, y, stride, tilingWidth }) {
   const arrSlice = slice(stride * tilingWidth * y + stride * x, stride * tilingWidth * y + stride * (x + 1));
-  const zarrKey = config.zarrConfig.store + config.zarrConfig.path;
-  if (!zarrArrays[zarrKey]) {
-    zarrArrays[zarrKey] = await openArray(config.zarrConfig);
-  }
-  const arr = zarrArrays[zarrKey];
   const dataSlice = await arr.get([arrSlice]);
   const { data } = dataSlice;
-  const { channelName } = config;
   const dataObj = {}
-  dataObj[channelName] = data
+  dataObj[channel] = data
   return dataObj;
 }
 
-export default function loadZarr({ sourceChannels, tileSize, x, y, z, imageWidth }) {
+export function loadZarr({ connections, tileSize, x, y, z, imageWidth }) {
   const tilingWidth = Math.ceil(imageWidth / (tileSize * (2 ** z)));
-  const dataNames = ['redData', 'greenData', 'blueData']; // denotes order rgb
-  const configList = Object.keys(sourceChannels).map((channel, i) => ({
-    channelName: channel,
-    channelType: dataNames[i],
-    zarrConfig: {
-      store: `${sourceChannels[channel]}/`,
-      path: `pyramid_${z}.zarr`,
-      mode: 'r',
-    },
-  }));
   const stride = tileSize * tileSize;
-  const configListPromises = configList.map((config) => {
+  const channelPromises = Object.keys(connections).map((channel) => {
     return getData({
-      config, tileSize, x, y, stride, tilingWidth,
+      arr: connections[channel][z], channel, tileSize, x, y, stride, tilingWidth,
     });
   });
-  return Promise.all(configListPromises).then((dataList) => {
+  return Promise.all(channelPromises).then((dataList) => {
     const orderedList = []
     const dataObj = Object.assign({}, ...dataList)
     Object.keys(dataObj).sort().forEach(function(key) {
@@ -43,4 +25,24 @@ export default function loadZarr({ sourceChannels, tileSize, x, y, z, imageWidth
     })
     return orderedList;
   });
+}
+
+export function getZarrConnections({sourceChannels, maxZoom}) {
+  const zarrConnections = Object.keys(sourceChannels).map(async (channel) => {
+    var zarrObj = {}
+    for(var i = 0; i < -maxZoom; i++) {
+      var zarr = await openArray({
+        store: `${sourceChannels[channel]}/`,
+        path: `pyramid_${i}.zarr`,
+        mode: 'r',
+      })
+      if(!zarrObj[channel]) {
+        zarrObj[channel] = [zarr]
+      } else {
+        zarrObj[channel].push(zarr)
+      }
+    }
+    return zarrObj
+  })
+  return Promise.all(zarrConnections)
 }
